finished trello clone

+++ flask-lessons/trello-clone/blueprints/cards_bp.py
@@ -2,10 +2,12 @@ from flask import Blueprint, request, abort
 from flask_jwt_extended import jwt_required, get_jwt_identity
 from setup import db
 from models.card import CardSchema, Card
-from auth import admin_required
+from auth import authorize
 from blueprints.comments_bp import comments_bp
 
 cards_bp = Blueprint('cards', __name__, url_prefix='/cards')
+
+cards_bp.register_blueprint(comments_bp)
 
 # Get all cards
 @cards_bp.route("/")
@@ -33,7 +35,6 @@ def one_card(id):
 @cards_bp.route('/', methods=['POST'])
 @jwt_required()
 def create_card():
-    admin_required()
     card_info = CardSchema(exclude=['id', 'date_created']).load(request.json)
     card = Card(
         title = card_info['title'],
@@ -49,11 +50,11 @@ def create_card():
 @cards_bp.route('/<int:id>', methods=['PUT', 'PATCH'])
 @jwt_required()
 def update_card(id):
-    admin_required()
     card_info = CardSchema(exclude=['id', 'date_created']).load(request.json)
     stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         card.title = card_info.get('title', card.title)
         card.description = card_info.get('description', card.description)
         card.status = card_info.get('status', card.status)
@@ -66,15 +67,14 @@ def update_card(id):
 @cards_bp.route('/<int:id>', methods=['DELETE'])
 @jwt_required()
 def delete_card(id):
-    admin_required()
     stmt = db.select(Card).filter_by(id=id) # .where(Card.id == id)
     card = db.session.scalar(stmt)
     if card:
+        authorize(card.user_id)
         db.session.delete(card)
         db.session.commit()
         return {}, 200
     else:
         return {'error': 'Card not found'}, 404
-    
 
-cards_bp.register_blueprint(comments_bp)
+
